[EDITORS] [REFACTO] Consolidate all editors under game-editor/editors
Move SpriteEditor and multiplayer from src/modules/editor into
src/modules/create/game-editor/editors, alongside MapEditor and SoundEditor,
and remove the now-empty src/modules/editor. References rewritten to the new
alias path.

Verified: typecheck, lint, and build all pass.
