+++ src/modules/create/game-editor/GameEditor.tsx
@@ -2,10 +2,10 @@ import { workSessionControllerLeave } from "@api";
 import CodeEditor from "@modules/create/game-editor/editors/CodeEditor";
 import GameEditorConsole from "@modules/create/game-editor/editors/GameEditorConsole";
 import { MapEditor } from "@modules/create/game-editor/editors/MapEditor/MapEditor";
+import { MultiplayerSettingsEditor } from "@modules/create/game-editor/editors/multiplayer/MultiplayerSettingsEditor";
 import ProjectSettingsEditor from "@modules/create/game-editor/editors/ProjectSettingsEditor";
 import { SoundEditor } from "@modules/create/game-editor/editors/SoundEditor/SoundEditor";
-import { MultiplayerSettingsEditor } from "@modules/editor/multiplayer/MultiplayerSettingsEditor";
-import { SpriteEditor } from "@modules/editor/SpriteEditor/SpriteEditor";
+import { SpriteEditor } from "@modules/create/game-editor/editors/SpriteEditor/SpriteEditor";
 import { ProjectProvider, ProviderEventType } from "@providers/ProjectProvider";
 import GameCanvas from "@shared/canvas/gameCanvas/GameCanvas";
 import { SpriteRendererHandle } from "@shared/canvas/RendererHandle";
